fix: DIPA-159 move gantt-menu to menus folder
