+++ frontend/projects/dipa-frontend/src/app/app.module.ts
@@ -3,7 +3,7 @@ import { LOCALE_ID, NgModule } from '@angular/core';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { GanttMenuComponent } from './modules/gantt/gantt-menu/gantt-menu.component';
+import { GanttMenuComponent } from './menus/gantt-menu/gantt-menu.component';
 import { ApiModule, Configuration } from 'dipa-api-client';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CoreModule } from './core/core.module';
