+++ frontend/projects/dipa-frontend/src/app/menus/gantt-menu/gantt-menu.component.html
